+++ src/main/kotlin/us/frollo/frollosdk/network/api/AggregationAPI.kt
@@ -24,7 +24,6 @@ import retrofit2.http.HTTP
 import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Path
-import retrofit2.http.Query
 import retrofit2.http.QueryMap
 import us.frollo.frollosdk.model.api.aggregation.accounts.AccountResponse
 import us.frollo.frollosdk.model.api.aggregation.accounts.AccountUpdateRequest
@@ -94,7 +93,7 @@ internal interface AggregationAPI {
     fun fetchProviderAccount(@Path("provider_account_id") providerAccId: Long): Call<ProviderAccountResponse>
 
     @PUT(URL_PROVIDER_ACCOUNTS)
-    fun refreshProviderAccounts(@Query("provideraccount_ids") providerAccIds: LongArray): Call<List<ProviderAccountResponse>>
+    fun refreshProviderAccounts(@QueryMap queryParams: Map<String, String>): Call<List<ProviderAccountResponse>>
 
     @POST(URL_PROVIDER_ACCOUNTS)
     fun createProviderAccount(@Body request: ProviderAccountCreateRequest): Call<ProviderAccountResponse>